adapt to server data. refactor.
