+++ src/components/place-card/place-card.jsx
@@ -4,12 +4,9 @@ import PropTypes from "prop-types";
 import {connect} from "react-redux";
 
 import {Settings} from "../../const";
-import {
-  debounce,
-  getRatingWidth
-} from "../../utils";
+import {debounce, getRatingWidth} from "../../utils";
 import {offerPropTypes} from "../../prop-types/offer";
-import {actionCreator} from "../../store/action";
+import {actionCreator} from "../../store/actions";
 
 export const PlaceCardComponent = (props) => {
   const {offer, classNameArticle, classNameWrapper, onActiveOfferChange} = props;
@@ -31,7 +28,7 @@ export const PlaceCardComponent = (props) => {
         <Link to={`/offer/${offer.id}`}>
           <img
             className="place-card__image"
-            src={offer.images[0]}
+            src={offer.previewImage}
             width="260"
             height="200"
             alt={offer.title}
