+++ src/components/property-review/property-review.jsx
@@ -12,10 +12,10 @@ export const PropertyReview = ({review}) => {
     <li className="reviews__item">
       <div className="reviews__user user">
         <div className="reviews__avatar-wrapper user__avatar-wrapper">
-          <img className="reviews__avatar user__avatar" src={review.avatar} width="54" height="54"
-            alt={`${review.name} avatar`}/>
+          <img className="reviews__avatar user__avatar" src={review.user.avatarUrl} width="54" height="54"
+            alt={`${review.user.name} avatar`}/>
         </div>
-        <span className="reviews__user-name">{review.name}</span>
+        <span className="reviews__user-name">{review.user.name}</span>
       </div>
       <div className="reviews__info">
         <div className="reviews__rating rating">
@@ -24,7 +24,7 @@ export const PropertyReview = ({review}) => {
             <span className="visually-hidden">Rating</span>
           </div>
         </div>
-        <p className="reviews__text">{review.review}</p>
+        <p className="reviews__text">{review.comment}</p>
         <time
           className="reviews__time"
           dateTime={getReviewTimeDateString(review.date)}
